+++ py-polars/src/interop/mod.rs
@@ -1,1 +1,2 @@
 pub mod arrow;
+pub mod numpy;
