+++ py-polars/src/lib.rs
@@ -36,7 +36,6 @@ mod py_modules;
 mod series;
 #[cfg(feature = "sql")]
 mod sql;
-mod to_numpy;
 mod utils;
 
 use pyo3::panic::PanicException;
